ollama_chat_rag/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
ollama
langchain
langchain-community
typer[all]
beautifulsoup4
unstructured
faiss-cpu
requests
